Movement Collisions
Used raycast, player size values are hardcoded for now.

If you try to normalize a vector using "normalized" it will return a vector that its value is normalized to 1. So if you only work with 1 dimension, x-y-z, that dimension value will return as 1 to normalize the total vector value to 1.

for Vector2 a, b = a(.5f, 1.5f, 0), b(.2f, 0, 0) when both vectors are normalized, both will have the value 1 but the values will be: for example a(.316f, .948f, 0), b(1f, 0, 0) - Exact output on wolframalpha.com