Game Manager
NOTE:
When using TextMeshPro or alike, don't directly modify the effects like outline, because it will modify the default material used by all related components.

Also the NAME of the material must include the name of the FONT or it will not show up in the editor
Example: "LiberationSans SDF StartCountDown"

GameManager with state machine to control the game flow.

Game start countdown added.