Sync kitchen object parenting (holding)
Continue from the last commit with the item sync.

MUST KNOW:
You cannot make a NetworkObject into a child of a Dynamically-Spawned NetworkObject.
And a NetworkObject cannot be a child of a non-NetworkObject parent.

So parenting don't really work on Multiplayer.

Here, we use parenting as an easy solution to movement/following,
so we can use a follow script instead.

On the KitchenGameMultiplayer script, after Intantiating and Spawning the network object, the rest of the code related to the parent assignment only works on Server + Host(if not a dedicated SV).

We need to make an Rpc for the parenting to work.

THIS IS A MUST IN MULTIPLAYER LOGIC
Understand that every non-static event or function will require Synchronization, especially if it is an interactive thing in the network.
The Server and the Client Rpc information and the setup must be absorbed well.

Link to the docs: https://docs-multiplayer.unity3d.com/netcode/current/basics/networkobject/

Currently Pickup and Drop works on clear counters, container counters, plate counters(Plates need revisiting).