Cutting Counter Sync
In the cutting counter interaction initial singleplayer code sets the kitchenobject's parent as itself. (Taking the object from the player)
But in mutliplayer this code uses a serverRpc, there will be a delay on server-client communications. Current code doesn't acknowledge this and executes the rest of the code normally, which later turns into a null reference exception problem. Because the object  on the player's hand won't immediately switch parents and become cutting counter's posession, the code doesn't get the object reference in time with the rest of the execution process.

Keep in mind, while the host executes the code as if it is a singleplayer game, client will have delay on some code executions because of networking.

Sync Info:
Changes that include the solutions:
- Interact and Parenting
- Alternative Interact
- Progress Syncing
- Decoupling Logic Tests -> Cutting Progress and Completion Checks