Stove Counter Sync
Again, because of the kitchenObject not switching parents in expected time during script execution, Parental Sync is required.
This can be easily done with caching the object we're switching parents of in our script.

Only the server should have the timer logic on the frying. Using Network Variable for it.

Same for the State and burnTimer
Also when giving the item to the player, state change should be a serverside change or the progressbar doesn't disappear.