Handling Late Joins
Link to full explanation of how to manage it: https://youtu.be/7glCsF9fv3s?si=cDcbgWHsnjxoaXmH&t=12022

Main problem is data desync. Network Variables on client are updated when a they join but the Rpc's that triggered before to update those Network Variables are not executed.

Which means that any code that was in the Rpc that changed the Network Variable's state, triggered events or logic didn't run. Causing data desync.

You cannot put Scriptable Objects inside Network Variables, thats why in our game Recipe Orders, Plate Counter etc. are using Rpc's to sync up.

You could use INetworkSerializable to override how to serialize the data you store in the scriptable objects etc. See network tutorial video of CodeMonkey.

You could listen to NetworkManager.Singleton.OnClientConnectedCallback to update logic on that specific client.

In this game we will not accept late joins.

We do need to call StartHost and Client functions of NetworkManager Manually to use the callback though.