Multiplayer Scene Management
Network Manager's setting: Enable Scene Management must be enabled. To sync scene loads on clients.

Then in Loader, instead of calling SceneManager.LoadScene we need NetworkManager.SceneManager because if you use the SceneManager directly, the objects on the new scene will not spawn.

Also players must be spawned manually now. Because the Network Manager is not in the GameScene. (Player Prefab on that will spawn a player, we had to disable it)